requirements.txt
absl-py==2.3.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
audioread==3.0.1
Bottleneck==1.5.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
coloredlogs==15.0.1
crcmod==1.7
cryptography==45.0.5
datasets==2.14.7
decorator==5.2.1
dill==0.3.7
Distance==0.1.3
edit-distance==1.0.6
editdistance==0.8.1
einops==0.8.1
espnet-tts-frontend==0.0.3
et_xmlfile==2.0.0
filelock==3.12.0
frozenlist==1.7.0
fsspec==2023.10.0
funasr==0.8.0
g2p==2.2.2
g2p-en==2.1.0
gast==0.6.0
grpcio==1.60.0
grpcio-tools==1.60.0
h5py==3.14.0
hdbscan==0.8.40
hf-xet==1.1.5
huggingface-hub==0.33.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
inflect==7.5.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.1
kaldiio==2.18.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8.2
MarkupSafe==3.0.2
modelscope==1.10.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.3
multiprocess==0.70.15
munkres==1.1.4
nara-wpe==0.0.11
networkx==3.5
nltk==3.9.1
numba==0.61.2
numpy==1.26.0
omegaconf==2.3.0
openpyxl==3.1.5
oss2==2.19.1
packaging==23.2
pandas==2.3.1
panphon==0.22.2
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==4.25.1
pyarrow==21.0.0
pyarrow-hotfix==0.7
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pynndescent==0.5.13
pypinyin==0.54.0
python-dateutil==2.9.0.post0
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
responses==0.18.0
rotary-embedding-torch==0.8.8
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentencepiece==0.2.0
simplejson==3.20.1
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
text-unidecode==1.3
TextGrid==1.6.1
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.1.0
torch-complex==0.4.4
torchaudio==2.1.0
tqdm==4.65.0
transformers==4.53.2
typeguard==4.4.4
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
unicodecsv==0.14.1
Unidecode==1.4.0
urllib3==2.5.0
webrtcvad==2.0.10
Werkzeug==3.1.3
xxhash==3.5.0
yapf==0.43.0
yarl==1.20.1
